requirements.txt
python-dotenv
langchain-community>=0.4,<0.5
langchain-core>=1.0,<2.0
langchain-google-genai>=4.0,<5.0
langchain-text-splitters>=1.0,<2.0
PyPDF2
faiss-cpu
streamlit
-e .
